refactor: move take to its own file
